pyproject.toml
[project]
name = "show-me-the-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "finance-datareader>=0.9.96",
    "google-generativeai>=0.8.5",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "streamlit>=1.28.0",
    "yfinance>=0.2.0",
]
